CMake fixes works with VS17 (just open the folder with the CMakeLists.txt and VS gets everything it needes)

+++ src/Camera.cpp
@@ -1,6 +1,6 @@
 #include "../include/Camera.h"
 #include <glm/gtc/matrix_transform.hpp>
-#include <glm/gtx/rotate_vector.hpp>;
+#include <glm/gtx/rotate_vector.hpp>
 
 void lpe::Camera::Copy(const Camera& other)
 {
@@ -67,6 +67,5 @@ glm::mat4 lpe::Camera::GetView() const
 
 glm::mat4 lpe::Camera::GetPerspective() const
 {
-  return glm::perspective(glm::radians(fov), swapChainExtent.width / (float)swapChainExtent.height, near, far);
-  //return glm::perspectiveFov(glm::radians(fov), (float)swapChainExtent.width, (float)swapChainExtent.height, near, far);
+  return glm::perspective(glm::radians(fov), static_cast<float>(swapChainExtent.width) / static_cast<float>(swapChainExtent.height), near, far);
 }
